Revert "Improved config. Fixes #9"
This reverts commit d948a957cbfd78b387fd2d9419d52019498a19ec.

+++ src/main/java/com/minecraftabnormals/allurement/core/mixin/ExperienceOrbEntityMixin.java
@@ -22,7 +22,7 @@ public class ExperienceOrbEntityMixin {
 	private void onCollideWithPlayer(PlayerEntity player, CallbackInfo ci) {
 		int count = AllurementUtil.getTotalEnchantmentLevel(AllurementEnchantments.ALLEVIATING.get(), player, EquipmentSlotType.Group.ARMOR);
 		if (count > 0) {
-			float factor = AllurementConfig.COMMON.alleviatingHealingFactor.get().floatValue() * count;
+			float factor = AllurementConfig.COMMON.alleviatingHealingFactor.get() * count;
 			float i = Math.min(this.xpValue * factor, player.getMaxHealth() - player.getHealth());
 			this.xpValue -= Math.round(i / factor);
 			player.heal(i);
